Addressing problem of split clusters between merged and non-merged reads

+++ results/2015-12-16b/README.sh
@@ -59,6 +59,8 @@ for j in 1 7 13 19; do
    done
    wait
 done
+# Clean up. I remove discarded reads.
+find merged -name '*.discarded.fastq' -exec rm '{}' \;
 
 # 2016-04-13: I just learned that PEAR has reverse-complemented second reads.
 # This has implications in downstream analysis (trimming of adapters, and pyrad).
@@ -109,27 +111,27 @@ wait
 # demultiplexing, because it gives more information in the stats. Note that now
 # I run this in the server, all samples at once.
 for i in `seq 1 24`; do
-   if [ ! -d demultiplexed/${SAMPLE[$i]} ]; then mkdir demultiplexed/${SAMPLE[$i]}; fi
-   if [ ! -d demultiplexed/${SAMPLE[$i]}/pe_full ]; then mkdir demultiplexed/${SAMPLE[$i]}/pe_full; fi
-   if [ ! -d demultiplexed/${SAMPLE[$i]}/pe_trunc ]; then mkdir demultiplexed/${SAMPLE[$i]}/pe_trunc; fi
-   if [ ! -d demultiplexed/${SAMPLE[$i]}/se_full ]; then mkdir demultiplexed/${SAMPLE[$i]}/se_full; fi
-   if [ ! -d demultiplexed/${SAMPLE[$i]}/se_trunc ]; then mkdir demultiplexed/${SAMPLE[$i]}/se_trunc; fi
-   if [ ! -e demultiplexed/${SAMPLE[$i]}/barcode.1 ]; then
+   if [ ! -d demultiplexed/${SAMPLE[$i]} ] && [ ! -e demultiplexed/${SAMPLE[$i]}.fastq ]; then mkdir demultiplexed/${SAMPLE[$i]}; fi
+   if [ ! -d demultiplexed/${SAMPLE[$i]}/pe_full ] && [ ! -e demultiplexed/${SAMPLE[$i]}_R1.fastq ]; then mkdir demultiplexed/${SAMPLE[$i]}/pe_full; fi
+   if [ ! -d demultiplexed/${SAMPLE[$i]}/pe_trunc ] && [ ! -e demultiplexed/${SAMPLE[$i]}_R1.fastq ]; then mkdir demultiplexed/${SAMPLE[$i]}/pe_trunc; fi
+   if [ ! -d demultiplexed/${SAMPLE[$i]}/se_full ] && [ ! -e demultiplexed/${SAMPLE[$i]}.fastq ]; then mkdir demultiplexed/${SAMPLE[$i]}/se_full; fi
+   if [ ! -d demultiplexed/${SAMPLE[$i]}/se_trunc ] && [ ! -e demultiplexed/${SAMPLE[$i]}.fastq ]; then mkdir demultiplexed/${SAMPLE[$i]}/se_trunc; fi
+   if [ ! -e demultiplexed/${SAMPLE[$i]}/barcode.1 ] && [ ! -e demultiplexed/${SAMPLE[$i]}.fastq ]; then
       echo -e ${SAMPLE[$i]}"\t"${BARCODE[$i]} > demultiplexed/${SAMPLE[$i]}/barcode.1
    fi
    # I notice that many reads miss the first base, and the truncated barcode is not
    # recognized. I will run pyrad's step 1 twice, once with the complete barcode and
    # 1 mismatch allowed, and once with the truncated barcode, without any mismatch allowed.
-   if [ ! -e demultiplexed/${SAMPLE[$i]}/barcode.2 ]; then
+   if [ ! -e demultiplexed/${SAMPLE[$i]}/barcode.2 ] && [ ! -e demultiplexed/${SAMPLE[$i]}.fastq ]; then
       gawk '{print $1 "\t" substr($2,2)}' demultiplexed/${SAMPLE[$i]}/barcode.1 > demultiplexed/${SAMPLE[$i]}/barcode.2
    fi
-   if [ ! -e demultiplexed/${SAMPLE[$i]}'_R1_.fastq' ]; then
-      ln -s `pwd`"/merged/${SAMPLE[$i]}.unassembled.forward.fastq" demultiplexed/${SAMPLE[$i]}'_R1_.fastq'
+   if [ ! -e demultiplexed/${SAMPLE[$i]}_R1_.fastq ] && [ ! -e demultiplexed/${SAMPLE[$i]}_R1.fastq ]; then
+      ln -s `pwd`"/merged/${SAMPLE[$i]}.unassembled.forward.fastq" demultiplexed/${SAMPLE[$i]}_R1_.fastq
    fi
-   if [ ! -e demultiplexed/${SAMPLE[$i]}'_R2_.fastq' ]; then
-      ln -s `pwd`"/merged/${SAMPLE[$i]}.unassembled.reverse.fastq" demultiplexed/${SAMPLE[$i]}'_R2_.fastq'
+   if [ ! -e demultiplexed/${SAMPLE[$i]}_R2_.fastq ] && [ ! -e demultiplexed/${SAMPLE[$i]}_R2.fastq ]; then
+      ln -s `pwd`"/merged/${SAMPLE[$i]}.unassembled.reverse.fastq" demultiplexed/${SAMPLE[$i]}_R2_.fastq
    fi
-   if [ ! -e demultiplexed/${SAMPLE[$i]}/params_1.txt ]; then
+   if [ ! -e demultiplexed/${SAMPLE[$i]}/params_1.txt ] && [ ! -e demultiplexed/${SAMPLE[$i]}_R1.fastq ]; then
       pyrad -n
       sed -i '/## 1. /c\demultiplexed/'${SAMPLE[$i]}'/pe_full/   ## 1. Working directory                                 (all)' params.txt
       sed -i '/## 2. /c\demultiplexed/'${SAMPLE[$i]}'*.fastq     ## 2. Loc. of non-demultiplexed files (if not line 18)  (s1)' params.txt
@@ -140,7 +142,7 @@ for i in `seq 1 24`; do
       sed -i '/## 19./c\1                                        ## 19.opt.: maxM: N mismatches in barcodes (def= 1)          (s1)' params.txt
       mv params.txt demultiplexed/${SAMPLE[$i]}/params_1.txt
    fi
-   if [ ! -e demultiplexed/${SAMPLE[$i]}/params_2.txt ]; then
+   if [ ! -e demultiplexed/${SAMPLE[$i]}/params_2.txt ] && [ ! -e demultiplexed/${SAMPLE[$i]}_R1.fastq ]; then
       pyrad -n
       sed -i '/## 1. /c\demultiplexed/'${SAMPLE[$i]}'/pe_trunc/  ## 1. Working directory                                 (all)' params.txt
       sed -i '/## 2. /c\demultiplexed/'${SAMPLE[$i]}'*.fastq     ## 2. Loc. of non-demultiplexed files (if not line 18)  (s1)' params.txt
@@ -151,7 +153,7 @@ for i in `seq 1 24`; do
       sed -i '/## 19./c\0                                        ## 19.opt.: maxM: N mismatches in barcodes (def= 1)          (s1)' params.txt
       mv params.txt demultiplexed/${SAMPLE[$i]}/params_2.txt
    fi
-   if [ ! -e demultiplexed/${SAMPLE[$i]}/params_3.txt ]; then
+   if [ ! -e demultiplexed/${SAMPLE[$i]}/params_3.txt ] && [ ! -e demultiplexed/${SAMPLE[$i]}.fastq ]; then
       pyrad -n
       sed -i '/## 1. /c\demultiplexed/'${SAMPLE[$i]}'/se_full/   ## 1. Working directory                                 (all)' params.txt
       sed -i '/## 2. /c\merged/'${SAMPLE[$i]}'.assembled.fastq   ## 2. Loc. of non-demultiplexed files (if not line 18)  (s1)' params.txt
@@ -162,7 +164,7 @@ for i in `seq 1 24`; do
       sed -i '/## 19./c\1                                        ## 19.opt.: maxM: N mismatches in barcodes (def= 1)          (s1)' params.txt
       mv params.txt demultiplexed/${SAMPLE[$i]}/params_3.txt
    fi
-   if [ ! -e demultiplexed/${SAMPLE[$i]}/params_4.txt ]; then
+   if [ ! -e demultiplexed/${SAMPLE[$i]}/params_4.txt ] && [ ! -e demultiplexed/${SAMPLE[$i]}.fastq ]; then
       pyrad -n
       sed -i '/## 1. /c\demultiplexed/'${SAMPLE[$i]}'/se_trunc/  ## 1. Working directory                                 (all)' params.txt
       sed -i '/## 2. /c\merged/'${SAMPLE[$i]}'.assembled.fastq   ## 2. Loc. of non-demultiplexed files (if not line 18)  (s1)' params.txt
@@ -173,16 +175,16 @@ for i in `seq 1 24`; do
       sed -i '/## 19./c\0                                        ## 19.opt.: maxM: N mismatches in barcodes (def= 1)          (s1)' params.txt
       mv params.txt demultiplexed/${SAMPLE[$i]}/params_4.txt
    fi
-   if [ ! -e demultiplexed/${SAMPLE[$i]}/pe_full/stats/s1.sorting.txt ]; then
+   if [ ! -e demultiplexed/${SAMPLE[$i]}/pe_full/stats/s1.sorting.txt ] && [ ! -e demultiplexed/${SAMPLE[$i]}_R1.fastq ]; then
       pyrad -p demultiplexed/${SAMPLE[$i]}/params_1.txt -s 1 &
    fi
-   if [ ! -e demultiplexed/${SAMPLE[$i]}/pe_trunc/stats/s1.sorting.txt ]; then
+   if [ ! -e demultiplexed/${SAMPLE[$i]}/pe_trunc/stats/s1.sorting.txt ] && [ ! -e demultiplexed/${SAMPLE[$i]}_R1.fastq ]; then
       pyrad -p demultiplexed/${SAMPLE[$i]}/params_2.txt -s 1 &
    fi
-   if [ ! -e demultiplexed/${SAMPLE[$i]}/se_full/stats/s1.sorting.txt ]; then
+   if [ ! -e demultiplexed/${SAMPLE[$i]}/se_full/stats/s1.sorting.txt ] && [ ! -e demultiplexed/${SAMPLE[$i]}.fastq ]; then
       pyrad -p demultiplexed/${SAMPLE[$i]}/params_3.txt -s 1 &
    fi
-   if [ ! -e demultiplexed/${SAMPLE[$i]}/se_trunc/stats/s1.sorting.txt ]; then
+   if [ ! -e demultiplexed/${SAMPLE[$i]}/se_trunc/stats/s1.sorting.txt ] && [ ! -e demultiplexed/${SAMPLE[$i]}.fastq ]; then
       pyrad -p demultiplexed/${SAMPLE[$i]}/params_4.txt -s 1 &
    fi
 done
@@ -220,10 +222,17 @@ for i in `seq 1 24`; do
    fi
 done
 wait
+# Clean up.
+for i in `seq 1 24`; do
+   if [ -e demultiplexed/${SAMPLE[$i]}_R1.fastq ] && [ -e demultiplexed/${SAMPLE[$i]}_R2.fastq ] && [ -e demultiplexed/${SAMPLE[$i]}.fastq ] && [ -d demultiplexed/${SAMPLE[$i]} ]; then
+      rm -r demultiplexed/${SAMPLE[$i]}
+      find demultiplexed -name '*_R1_*' -exec rm '{}' \;
+      find demultiplexed -name '*_R2_*' -exec rm '{}' \;
+   fi
+done
 # Now, we trim assembled and unassembled reads separately as well. Hopefully,
 # assembled reads should not need much trimming.
-for j in 1 4 7 10 13 16 19 22; do
-   for i in `seq $j $(( $j + 2 ))`; do
+   for i in `seq 1 24`; do
       if [ ! -e trimmed/${SAMPLE[$i]}_R1.fastq ]; then
          BARCODEREV=`echo ${BARCODE[$i]} | gawk -f revcomp.awk`
          cutadapt -a CGAGATCGGAAGAGCACACGTCTGAACTCCAGTCAC \
@@ -242,7 +251,6 @@ for j in 1 4 7 10 13 16 19 22; do
       fi
    done
    wait
-done
 
 ## Clean up.
 #for i in `seq 1 24`; do
